Module 10 - completed part 2 and part 3

+++ todo_app/app.py
@@ -3,17 +3,20 @@ from .data.todoapi import TodoAPI
 from .data.item import Item
 from dateutil import parser
 from .viewmodel import ViewModel
+from .useradminviewmodel import UserAdminViewModel
 from .app_config import Config
-from flask_login import LoginManager, login_required, login_user, logout_user
+from flask_login import LoginManager, login_required, login_user, logout_user, current_user
 from oauthlib.oauth2 import WebApplicationClient
 import requests
 from .data.user import User
+from .data.userdb import UserDB
 
 def create_app():
     app = Flask(__name__)
     app.config.from_object(Config()) 
     
     todoapi = TodoAPI(app.config) 
+    users_db = UserDB(app.config)
 
     login_manager = LoginManager()
     client = WebApplicationClient(app.config.get('GITHUB_CLIENT_ID', ''))
@@ -26,14 +29,28 @@ def create_app():
     
     @login_manager.user_loader
     def load_user(user_id):
-         return User(user_id)
+         return get_user(user_id)
     
     login_manager.init_app(app)    
     # All the routes and setup code etc
+    def get_user(user_id):
+        return users_db.get_user(user_id)
+    
+
+    def add_user(user_profile):
+        default_roles = ['reader']
+        if len(users_db.get_list_of_users()) == 0:
+            default_roles = ['admin']
+        
+        users_db.add_user({**user_profile, 'roles': default_roles})
     
     def render_index_response():
-        item_view_model = ViewModel(sorted(todoapi.get_list_of_items(), key=lambda item : (item.status, item.id)))
+        item_view_model = ViewModel(sorted(todoapi.get_list_of_items(), key=lambda item : (item.status, item.id)), current_user)
         return render_template('index.html', view_model= item_view_model)
+
+    def render_usersadmin_response():
+        useradmin_view_model = UserAdminViewModel(users_db.get_list_of_users(), current_user)
+        return render_template('useradmin.html', view_model= useradmin_view_model)
 
     @app.route('/login/callback', methods=['GET', 'POST'])
     def _callback():
@@ -45,7 +62,12 @@ def create_app():
         userinfo_endpoint = 'https://api.github.com/user'
         uri, headers, body = client.add_token(userinfo_endpoint)
         userinfo_response = requests.get(uri, headers=headers, data=body)
-        user = User(userinfo_response.json()['login'])
+        user_id = userinfo_response.json()['login']
+
+        if get_user(user_id) is None:
+            add_user(userinfo_response.json())
+        
+        user = get_user(user_id)
         login_user(user)
         
         return redirect('/', code=302)
@@ -56,6 +78,30 @@ def create_app():
         logout_user()
         return redirect('/', code=302)
 
+    @app.route('/useradmin', methods=['GET'])
+    @login_required
+    def _useradmin():
+        if 'admin' not in current_user.roles:
+            return redirect('/', code=302)
+        
+        return render_usersadmin_response()
+
+    @app.route('/usersetrole/<id>/<role>/<add_remove>', methods=['POST'])
+    @login_required
+    def _usersetrole(id,role,add_remove):
+        if 'admin' not in current_user.roles:
+            return redirect('/', code=302)
+        
+        roles = current_user.roles
+        if role in roles and add_remove == 'remove':
+            roles.remove(role)
+        elif role not in roles and add_remove == 'add':
+            roles.append(role)
+        else:
+            pass
+
+        users_db.modify_roles(id,roles)
+        return render_usersadmin_response()
 
     @app.route('/', methods=['GET'])
     @login_required
@@ -65,19 +111,22 @@ def create_app():
     @app.route('/add', methods=['POST'])
     @login_required
     def _add():
-        todoapi.add_item(request.form)
+        if 'writer' in current_user.roles:
+            todoapi.add_item(request.form)
         return render_index_response()
 
     @app.route('/setstatus/<id>/<status>', methods=['POST'])
     @login_required
     def _setstatus(id,status):
-        todoapi.modify_item(id, {'status': status})
+        if 'writer' in current_user.roles:
+            todoapi.modify_item(id, {'status': status})
         return render_index_response()
 
     @app.route('/delete/<id>', methods=['POST'])
     @login_required
     def _delete(id):
-        todoapi.delete_item(id)
+        if 'writer' in current_user.roles:
+            todoapi.delete_item(id)
         return render_index_response()
 
     @app.template_filter('strftime')
